requirements.txt
airflow-code-editor
apache-airflow[slack]